+++ src/main.cpp
@@ -137,7 +137,7 @@ static void init(const std::string render_device_name) {
 
 static void mainLoop (bool debug_event) {
 	bool done = false;
-	int delay = floor((1000.f/MAX_FRAMES_PER_SEC)+0.5f);
+	int delay = int(floor((1000.f/MAX_FRAMES_PER_SEC)+0.5f));
 	int logic_ticks = SDL_GetTicks();
 	int loops = 0;
